refactor: extract remaining AI planner hosts into single-concern siblings (#4530)
Split Full AI setup/merge, labour input types, lock-recovery seller flags, and strategic goal/army prep so the 300-line gate measures code rather than mixed hosts.
