+++ packages/colonizethis_ai/lib/src/planning/economy_planner_labour.dart
@@ -4,6 +4,7 @@ import 'ai_commodity_ids.dart';
 import 'economy_planner_constants.dart';
 import 'economy_planner_labour_fabric_prepass.dart';
 import 'economy_planner_labour_feedstock.dart';
+import 'economy_planner_labour_input.dart';
 import 'growth_stage.dart';
 import 'planning_imports.dart';
 import 'recipe_scoring.dart';
@@ -11,6 +12,8 @@ import 'scored_candidate.dart';
 
 export 'economy_planner_labour_feedstock.dart'
     show multiInputImprovementOutputs;
+export 'economy_planner_labour_input.dart'
+    show LabourAllocationInput, missingCheapestRegimentBuildInputIds;
 
 final _log = packageLogger('economy_planner_labour');
 
@@ -33,52 +36,6 @@ IndustryCounselGrowthStage? _industryCounselGrowthStage(GrowthStage? stage) {
     resourceProductionPriority: stage.resourceProductionPriority,
     militaryPriority: stage.militaryPriority,
   );
-}
-
-/// Bundles inputs for [allocateLabour] (Refs #3977 AC5).
-final class LabourAllocationInput {
-  const LabourAllocationInput({
-    required this.stockpile,
-    required this.workers,
-    required this.effectiveLabour,
-    required this.config,
-    required this.seeds,
-    this.techUnlocked,
-    this.militaryRebuildCrisis = false,
-    this.regimentBuildInputProductionBoost = false,
-    this.missingRegimentBuildInputIds = const {},
-    this.supplierReleaseImprovementInputIds = const {},
-    this.feedstockReserveOutputIds = const {},
-    this.castIronLabourPeasantRecruitFabricBoost = false,
-    this.growthStage,
-  });
-
-  final Stockpile stockpile;
-  final WorkerPool workers;
-  final int effectiveLabour;
-  final AIConfig config;
-  final AISeedBundle seeds;
-  final Map<String, bool>? techUnlocked;
-  final bool militaryRebuildCrisis;
-  final bool regimentBuildInputProductionBoost;
-  final Set<String> missingRegimentBuildInputIds;
-  final Set<String> supplierReleaseImprovementInputIds;
-  final Set<String> feedstockReserveOutputIds;
-  final bool castIronLabourPeasantRecruitFabricBoost;
-  final GrowthStage? growthStage;
-}
-
-/// Commodity ids the cheapest regiment still needs in the stockpile before
-/// `suggestBuildOrders` will surface it (Refs #2847 H8).
-Set<String> missingCheapestRegimentBuildInputIds(Stockpile stockpile) {
-  final missing = <String>{};
-  for (final entry
-      in RegimentEconomyCatalog.peasantLevies.buildInputs.entries) {
-    if (stockpile.quantityOf(entry.key) < entry.value) {
-      missing.add(entry.key);
-    }
-  }
-  return missing;
 }
 
 List<AssignedRecipe> allocateLabour(LabourAllocationInput input) {
@@ -123,7 +80,9 @@ List<AssignedRecipe> allocateLabour(LabourAllocationInput input) {
   // recipes the GP is actively producing (Refs #2847 H8-extraction feedstock
   // co-availability). Empty when no such recipe is targeted, in which case
   // feasibility falls back to the unreduced stockpile (behaviour-equal).
-  final feedstockReserve = feedstockReserveForOutputs(feedstockReserveOutputIds);
+  final feedstockReserve = feedstockReserveForOutputs(
+    feedstockReserveOutputIds,
+  );
   final labourByRecipe = <String, int>{};
 
   if (castIronLabourPeasantRecruitFabricBoost) {
